Lets make everything require being verified.

+++ cogs/mod_mail.py
@@ -2,6 +2,7 @@ import config
 import discord
 from discord.ext import commands
 from discord.ext.commands import Cog
+from helpers.checks import check_if_verified
 from helpers.userlogs import get_blank_userlog, get_userlog, set_userlog
 import json
 import time
@@ -48,12 +49,14 @@ class ModMail(Cog):
 
     # Commands
 
+    @commands.check(check_if_verified)
     @commands.command(aliases=["creport"])
     async def modmail(self, ctx, *, body: str = ""):
         """Sends a mod mail message"""
 
         # We should probably delete the message for privacy.
-        await ctx.message.delete()
+        if ctx.guild:
+            await ctx.message.delete()
 
         # Prevent sending of blank messages.
         if len(body.strip()) == 0:
@@ -107,6 +110,7 @@ class ModMail(Cog):
 
         await ctx.send(f"{ctx.author.mention} - Message sent.")
 
+    @commands.check(check_if_verified)
     @commands.command(aliases=["solved", "completed"])
     async def resolved(self, ctx):
         """Marks your last mod mail message as resolved"""
